Renamed PlatformUtil.kt to Platforms.kt in line with the kotlin spec.
